templates/mpesa-django-express/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
phonenumbers==8.13.35
psycopg2==2.9.9
python-decouple==3.8
requests==2.31.0